refactor: トランザクション管理を移動
