+++ internal/usecase/work.go
@@ -48,20 +48,6 @@ func (uc *WorkUseCase) CreateWork(ctx context.Context, title, description, descr
 	ctx, cancel := context.WithTimeout(ctx, uc.timeout)
 	defer cancel()
 
-	workRepoTx, err := uc.repo.BeginTx(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to begin transaction: %w", err)
-	}
-
-	defer func() {
-		if r := recover(); r != nil {
-			workRepoTx.Rollback()
-			panic(r)
-		} else if err != nil {
-			workRepoTx.Rollback()
-		}
-	}()
-
 	work := &entity.Work{
 		ID:              uuid.New(),
 		Title:           title,
@@ -75,15 +61,9 @@ func (uc *WorkUseCase) CreateWork(ctx context.Context, title, description, descr
 		return nil, fmt.Errorf("validation failed: %w", err)
 	}
 
-	createdWork, err := workRepoTx.Create(ctx, work)
+	createdWork, err := uc.repo.Create(ctx, work)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create work: %w", err)
 	}
-
-	err = workRepoTx.Commit()
-	if err != nil {
-		return nil, fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
 	return createdWork, nil
 }
